add support for AC5’s ULZ compression
Translated the ULZ decompression code of my “unpack_AC3” program from C++ to C#. The ULZ used in AC5 seems to be identical to the one used in AC3.

The bigger challenge was storing the extracted data in the tree view nodes because it was never designed for displaying data that cannot be found directly in the file.